+++ src/main/java/cc/cassian/item_descriptions/client/jade/JadeBlockDescriptions.java
@@ -1,6 +1,8 @@
 package cc.cassian.item_descriptions.client.jade;
 
 import cc.cassian.item_descriptions.client.ModClient;
+import cc.cassian.item_descriptions.client.config.ModConfig;
+import cc.cassian.item_descriptions.client.helpers.TagHelpers;
 import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 import snownee.jade.api.BlockAccessor;
@@ -20,7 +22,13 @@ public enum JadeBlockDescriptions implements IBlockComponentProvider {
         //Check if block descriptions are enabled in mod config.
         if (showBlockDescriptions()) {
             //Create and add tooltip.
-            List<Text> tooltip = createTooltip(createBlockDescription(blockAccessor.getBlock(), blockAccessor.getLevel(), blockAccessor.getPosition(), blockAccessor.getBlockState(), blockAccessor.getBlockEntity()), true);
+            List<Text> tooltip;
+            if (ModConfig.get().developer_showAllPotentialKeys) {
+                tooltip = TagHelpers.findAllPotentialKeys(blockAccessor.getBlockState());
+            }
+            else {
+                tooltip = createTooltip(createBlockDescription(blockAccessor.getBlock(), blockAccessor.getLevel(), blockAccessor.getPosition(), blockAccessor.getBlockState(), blockAccessor.getBlockEntity()), true);
+            }
             for (Text text : tooltip) {
                 lines.add(text);
             }