+++ src/main/java/cc/cassian/item_descriptions/client/wthit/WTHITIntegration.java
@@ -1,5 +1,7 @@
 package cc.cassian.item_descriptions.client.wthit;
 
+import cc.cassian.item_descriptions.client.config.ModConfig;
+import cc.cassian.item_descriptions.client.helpers.TagHelpers;
 import mcp.mobius.waila.api.*;
 import net.minecraft.block.Block;
 import net.minecraft.entity.Entity;
@@ -28,7 +30,13 @@ public class WTHITIntegration implements IWailaPlugin, IBlockComponentProvider, 
     public void appendBody(ITooltip lines, IBlockAccessor blockAccessor, IPluginConfig config) {
         //Check if block descriptions are enabled in mod config.
         if (showBlockDescriptions() && config.getBoolean(BLOCK_DESCRIPTIONS)) {
-            List<Text> tooltip = createTooltip(createBlockDescription(blockAccessor.getBlock(), blockAccessor.getWorld(), blockAccessor.getPosition(), blockAccessor.getBlockState(), blockAccessor.getBlockEntity()), true);
+            List<Text> tooltip;
+            if (ModConfig.get().developer_showAllPotentialKeys) {
+                tooltip = TagHelpers.findAllPotentialKeys(blockAccessor.getBlockState());
+            }
+            else {
+                tooltip = createTooltip(createBlockDescription(blockAccessor.getBlock(), blockAccessor.getWorld(), blockAccessor.getPosition(), blockAccessor.getBlockState(), blockAccessor.getBlockEntity()), true);
+            }
             for (Text text : tooltip) {
                 lines.addLine(text);
             }
@@ -37,9 +45,15 @@ public class WTHITIntegration implements IWailaPlugin, IBlockComponentProvider, 
 
     @Override
     public void appendBody(ITooltip lines, IEntityAccessor entityAccessor, IPluginConfig config) {
-        //Check if block descriptions are enabled in mod config.
+        //Check if entity descriptions are enabled in mod config.
         if (showEntityDescriptions()  && config.getBoolean(ENTITY_DESCRIPTIONS)) {
-            var tooltip = createEntityDescription(entityAccessor.getEntity());
+            List<Text> tooltip;
+            if (ModConfig.get().developer_showAllPotentialKeys) {
+                tooltip = TagHelpers.findAllPotentialKeys(entityAccessor.getEntity());
+            }
+            else {
+                tooltip = createEntityDescription(entityAccessor.getEntity());
+            }
             for (Text text : tooltip) {
                 lines.addLine(text);
             }